BackEnd/auth_service/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
requests
passlib[bcrypt]
pytest
httpx
psycopg2-binary
python-jose
python-multipart
bcrypt
pydantic[email]>=2.0.0
pydantic-settings